refactor: centralize dummy db helpers
